# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Web API Dev Test - REST API Documentation",
    "version": "1.0.0",
    "description": "API documentation for server.js serving Sri Lanka provinces, districts, police stations, vehicles, and GPS pings."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Swagger Proxy Server (Bypasses CORS)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Direct API Server (server.js)"
    }
  ],
  "components": {
    "schemas": {
      "Province": {
        "type": "object",
        "properties": {
          "province_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Western Province"
          }
        }
      },
      "District": {
        "type": "object",
        "properties": {
          "district_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Colombo"
          },
          "province_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Station": {
        "type": "object",
        "properties": {
          "station_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Colombo Police Station"
          },
          "district_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "vehicle_id": {
            "type": "integer",
            "example": 1
          },
          "reg_number": {
            "type": "string",
            "example": "HB-6168"
          },
          "device_id": {
            "type": "string",
            "example": "TUK-DEV-520651"
          },
          "station_id": {
            "type": "integer",
            "example": 4
          }
        }
      },
      "VehicleComposite": {
        "type": "object",
        "properties": {
          "vehicle_id": {
            "type": "integer",
            "example": 1
          },
          "reg_number": {
            "type": "string",
            "example": "HB-6168"
          },
          "device_id": {
            "type": "string",
            "example": "TUK-DEV-520651"
          },
          "station_id": {
            "type": "integer",
            "example": 4
          },
          "last_ping": {
            "$ref": "#/components/schemas/Ping",
            "nullable": true
          }
        }
      },
      "LastPosition": {
        "type": "object",
        "properties": {
          "vehicle_id": {
            "type": "integer",
            "example": 1
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-14T00:00:00Z"
          },
          "lat": {
            "type": "number",
            "example": 7.312694
          },
          "lng": {
            "type": "number",
            "example": 80.60383
          },
          "speed": {
            "type": "number",
            "example": 45.5
          }
        }
      },
      "Ping": {
        "type": "object",
        "properties": {
          "ping_id": {
            "type": "integer",
            "example": 1
          },
          "vehicle_id": {
            "type": "integer",
            "example": 1
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-14T00:00:00Z"
          },
          "lat": {
            "type": "number",
            "example": 7.312694
          },
          "lng": {
            "type": "number",
            "example": 80.60383
          },
          "speed": {
            "type": "number",
            "example": 45.5
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found"
          }
        }
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Health check endpoint",
        "description": "Returns API status and student session ID.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "session": {
                      "type": "string",
                      "example": "NB6007CEM"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/provinces": {
      "get": {
        "summary": "Get all provinces",
        "description": "Retrieves a list of all provinces.",
        "responses": {
          "200": {
            "description": "A list of provinces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Province"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/provinces/{id}": {
      "get": {
        "summary": "Get province by ID",
        "description": "Retrieves details of a specific province by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the province to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Province details found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Province"
                }
              }
            }
          },
          "404": {
            "description": "Province not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/districts": {
      "get": {
        "summary": "Get all districts",
        "description": "Retrieves a list of all districts.",
        "responses": {
          "200": {
            "description": "A list of districts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/District"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/districts/{id}": {
      "get": {
        "summary": "Get district by ID",
        "description": "Retrieves details of a specific district by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the district to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "District details found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/District"
                }
              }
            }
          },
          "404": {
            "description": "District not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/stations": {
      "get": {
        "summary": "Get all police stations",
        "description": "Retrieves a list of all police stations.",
        "responses": {
          "200": {
            "description": "A list of police stations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stations/{id}": {
      "get": {
        "summary": "Get station by ID",
        "description": "Retrieves details of a specific police station by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the station to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Station details found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles": {
      "get": {
        "summary": "Get all vehicles",
        "description": "Retrieves a list of all vehicles along with their latest GPS ping.",
        "responses": {
          "200": {
            "description": "A list of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}": {
      "get": {
        "summary": "Get vehicle by ID",
        "description": "Retrieves details of a specific vehicle along with its latest GPS ping.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the vehicle to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle details found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleComposite"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/last-position": {
      "get": {
        "summary": "Get most recent position for a vehicle",
        "description": "Retrieves the most recent LocationPing coordinates for a specific vehicle.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the vehicle",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest position found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LastPosition"
                }
              }
            }
          },
          "404": {
            "description": "Position or vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/pings": {
      "get": {
        "summary": "Get ping history for a vehicle",
        "description": "Retrieves all GPS ping records associated with a specific vehicle ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the vehicle whose pings to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ping records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ping"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}